Improved Home Page Code
