--- code/public/CarsImagCards.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ca3fefbf6cc8f1f294c43786ed738dda1d766d702afc73c156944db2502558bf
-size 308173
