+++ code/tailwind.config.js
@@ -19,6 +19,7 @@ module.exports = {
         240: '60rem',
         213: '53.25rem',
         160: '40rem',
+        130: '32.5rem',
       },
       width: {
         1066: '266.5rem',
